Remove old CallInfo and fix some format
